+++ Aelia's Apprentice/Assets/Scripts/InteractableObject.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-public class InteractableObject : MonoBehaviour, IInteractable
-{
-    public void OnInteract()
-    {
-        Debug.Log(gameObject + ": I have been interacted with");
-    }
-}